requirements.txt
# ============================================================
# Pet Store SCM – MCP + LLM + CatBoost Forecasting
# Compatible with Python 3.10 / 3.11 / 3.12
# ============================================================

# --- Core ---
numpy>=1.26.0,<2.0
pandas>=2.1.0
python-dotenv>=1.0.0

# --- ML Forecasting (TFT primary, CatBoost fallback) ---
# PyTorch: install manually with CUDA for your system:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
pytorch-forecasting>=1.0.0   # TFT implementation
lightning>=2.0.0              # PyTorch Lightning trainer
catboost>=1.2.0               # CatBoost fallback (always available)
scikit-learn>=1.3.0           # metrics / preprocessing utilities
joblib>=1.3.0                 # model serialisation (save/load .pkl cache)

# --- Web Search (agent tools) ---
google-search-results>=2.4.2   # SerpAPI Google Search (set SERPAPI_KEY in .env)
ddgs>=0.1.0                    # Free DuckDuckGo fallback search (no API key needed)

# --- LLM Providers (all optional: install what you use) ---
anthropic>=0.49.0
openai>=1.30.0
groq>=0.11.0                   # Groq fast inference (uses OpenAI-compatible SDK)
google-generativeai>=0.8.0

# --- MCP Server ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0

# --- Database Drivers (optional: install if using live DB) ---
aiomysql>=0.2.0
asyncpg>=0.29.0

# --- UI ---
gradio>=5.0.0
plotly>=5.22.0

# --- Utilities ---
pydantic>=2.7.0
